initial split by dependency on cosmo, GCData, both, and neither
