+++ clmm/__init__.py
@@ -1,10 +1,13 @@
 """ CLMM is a cluster mass modeling code. """
 from .gcdata import GCData
 from .galaxycluster import GalaxyCluster
+
+from .dataops import compute_tangential_and_cross_components
 from .hybrid import make_binned_profile, convert_units
-from .polaraveraging import compute_tangential_and_cross_components
-from .utils import compute_radial_averages, make_bins
-from .modeling import cclify_astropy_cosmo, get_reduced_shear_from_convergence, get_3d_density, predict_surface_density, predict_excess_surface_density, angular_diameter_dist_a1a2, get_critical_surface_density, predict_tangential_shear, predict_convergence, predict_reduced_tangential_shear
+from .modeling import cclify_astropy_cosmo, get_3d_density, predict_surface_density, predict_excess_surface_density, angular_diameter_dist_a1a2, get_critical_surface_density, predict_tangential_shear, predict_convergence, predict_reduced_tangential_shear, astropyify_ccl_cosmo, predict_magnification
+# from .modeling import _convert_rad_to_mpc
+from .utils import compute_radial_averages, make_bins, convert_shapes_to_epsilon, build_ellipticities, compute_lensed_ellipticity, get_reduced_shear_from_convergence
+# from .utils import _get_a_from_z, _get_z_from_a, _compute_tangential_shear, _compute_cross_shear, _compute_lensing_angles_flatsky
 
 from . import lsst
