+++ clmm/dataops.py
@@ -1,22 +1,21 @@
-"""Functions to compute polar/azimuthal averages of data in radial bins"""
+"""Functions to compute polar/azimuthal averages of data in radial bins (formerly known as polaraveraging.py)"""
 # try: # 7481794
 #     import pyccl as ccl
 # except ImportError:
 #     pass
-import math
-import warnings
 import numpy as np
-from .gcdata import GCData
-from .utils import compute_radial_averages, make_bins, convert_units
+import warnings
+
 from .galaxycluster import GalaxyCluster
+from .gcdata import GCData
+from .utils import compute_radial_averages, make_bins, convert_units, _compute_tangential_shear, _compute_cross_shear
+
+from .hybrid import make_binned_profile
+from .modeling import cclify_astropy_cosmo, angular_diameter_dist_a1a2, predict_tangential_shear, predict_convergence, predict_reduced_tangential_shear, astropyify_ccl_cosmo, predict_magnification
+from .utils import compute_radial_averages, make_bins, convert_shapes_to_epsilon, build_ellipticities, compute_lensed_ellipticity, get_reduced_shear_from_convergence
 
 
-def compute_tangential_and_cross_components(cluster=None,
-                  shape_component1='e1', shape_component2='e2',
-                  tan_component='et', cross_component='ex',
-                  ra_lens=None, dec_lens=None, ra_source_list=None,
-                  dec_source_list=None, shear1=None, shear2=None, geometry='flat',
-                  add_to_cluster=True):
+def compute_tangential_and_cross_components(cluster=None, shape_component1='e1', shape_component2='e2', tan_component='et', cross_component='ex', ra_lens=None, dec_lens=None, ra_source_list=None, dec_source_list=None, shear1=None, shear2=None, geometry='flat', add_to_cluster=True):
     r"""Computes tangential- and cross- components for shear or ellipticity
 
     To compute the shear, we need the right ascension and declination of the lens and of
@@ -31,7 +30,7 @@ def compute_tangential_and_cross_components(cluster=None,
     2. Given a `GalaxyCluster` object::
 
         compute_shear(cluster)
-
+_compute_cross_shear
     3. As a method of `GalaxyCluster`::
 
         cluster.compute_shear()
@@ -76,7 +75,7 @@ def compute_tangential_and_cross_components(cluster=None,
     tan_component: string, optional
         Name of the column to be added to the `galcat` astropy table that will contain the
         tangential component computed from columns `shape_component1` and `shape_component2`.
-        Default: `et`
+        Default: `et`_compute_cross_shear
     cross_component: string, optional
         Name of the column to be added to the `galcat` astropy table that will contain the
         cross component computed from columns `shape_component1` and `shape_component2`.
@@ -149,76 +148,6 @@ def compute_tangential_and_cross_components(cluster=None,
         cluster.galcat[cross_component] = cross_shear
 
     return angsep, tangential_shear, cross_shear
-
-
-def _compute_lensing_angles_flatsky(ra_lens, dec_lens, ra_source_list, dec_source_list):
-    r"""Compute the angular separation between the lens and the source and the azimuthal
-    angle from the lens to the source in radians.
-
-    In the flat sky approximation, these angles are calculated using
-    .. math::
-        \theta = \sqrt{\left(\delta_s - \delta_l\right)^2 +
-        \left(\alpha_l-\alpha_s\right)^2\cos^2(\delta_l)}
-
-        \tan\phi = \frac{\delta_s - \delta_l}{\left(\alpha_l - \alpha_s\right)\cos(\delta_l)}
-
-    For extended descriptions of parameters, see `compute_shear()` documentation.
-    """
-    if not -360. <= ra_lens <= 360.:
-        raise ValueError(f"ra = {ra_lens} of lens if out of domain")
-    if not -90. <= dec_lens <= 90.:
-        raise ValueError(f"dec = {dec_lens} of lens if out of domain")
-    if not all(-360. <= x_ <= 360. for x_ in ra_source_list):
-        raise ValueError("Cluster has an invalid ra in source catalog")
-    if not all(-90. <= x_ <= 90 for x_ in dec_source_list):
-        raise ValueError("Cluster has an invalid dec in the source catalog")
-
-    # Put angles between -pi and pi
-    r2pi = lambda x: x - np.round(x/(2.0*math.pi))*2.0*math.pi
-
-    deltax = r2pi (np.radians(ra_source_list - ra_lens)) * math.cos(math.radians(dec_lens))
-    deltay = np.radians(dec_source_list - dec_lens)
-
-    # Ensure that abs(delta ra) < pi
-    #deltax[deltax >= np.pi] = deltax[deltax >= np.pi] - 2.*np.pi
-    #deltax[deltax < -np.pi] = deltax[deltax < -np.pi] + 2.*np.pi
-
-    angsep = np.sqrt(deltax**2 + deltay**2)
-    phi = np.arctan2(deltay, -deltax)
-    # Forcing phi to be zero everytime angsep is zero. This is necessary due to arctan2 features (it returns ).
-    phi[angsep==0.0] = 0.0
-
-    if np.any(angsep > np.pi/180.):
-        warnings.warn("Using the flat-sky approximation with separations >1 deg may be inaccurate")
-
-    return angsep, phi
-
-
-def _compute_tangential_shear(shear1, shear2, phi):
-    r"""Compute the tangential shear given the two shears and azimuthal positions for
-    a single source or list of sources.
-
-    We compute the tangential shear following Eq. 7 of Schrabback et al. 2018, arXiv:1611:03866
-    .. math::
-        g_t = -\left( g_1\cos\left(2\phi\right) - g_2\sin\left(2\phi\right)\right)
-
-    For extended descriptions of parameters, see `compute_shear()` documentation.
-    """
-    return - (shear1 * np.cos(2.*phi) + shear2 * np.sin(2.*phi))
-
-
-def _compute_cross_shear(shear1, shear2, phi):
-    r"""Compute the cross shear given the two shears and azimuthal position for a single
-    source of list of sources.
-
-    We compute the cross shear following Eq. 8 of Schrabback et al. 2018, arXiv:1611:03866
-    also checked arxiv 0509252
-    .. math::
-        g_x = g_1 \sin\left(2\phi\right) - g_2\cos\left(2\phi\right)
-
-    For extended descriptions of parameters, see `compute_shear()` documentation.
-    """
-    return shear1 * np.sin(2.*phi) - shear2 * np.cos(2.*phi)
 
 
 # Monkey patch functions onto Galaxy Cluster object
